move: dto-etc 폴더로 파일 이동 및 분리
- 요청/응답 dto가 아닌 기타 dto들을 모아놓기 위함
